requirements.txt
biopython==1.86
boto3==1.40.66
botocore==1.40.66
bottle==0.13.4
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
cheroot==11.1.1
ete4==4.4.0
idna==3.11
jaraco.functools==4.3.0
jmespath==1.0.1
more-itertools==10.8.0
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.14.0
scipy==1.16.3
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
